Conflicts:
	src/ui/hud.h

+++ src/ui/hud.h
@@ -57,7 +57,7 @@ protected:
     std::map<char, std::unique_ptr<VoxelCluster>> m_font3x5;
     std::map<char, std::unique_ptr<VoxelCluster>> m_font5x7;
     std::unique_ptr<HUDElement> m_shiparrow;
-    float m_delta_sec_remain;
+    double m_delta_sec_remain;
     float m_framerate;
     float m_dx, m_dy;
